#1607 Rename react-node sample app to be react-nodes-carbon
