+++ canvas_modules/harness/src/client/components/custom-canvases/react-nodes-carbon/react-nodes-carbon-palette.json
