+++ canvas_modules/harness/src/client/components/custom-canvases/react-nodes-carbon/react-nodes-carbon.scss
@@ -15,7 +15,7 @@
  */
 
 /* Override styles for react-nodes sample application. */
-.react-nodes {
+.react-nodes-carbon {
 
 	.d3-node-group {
 		.card-node-outline-div {
