+++ canvas_modules/harness/src/client/components/custom-canvases/react-nodes-carbon/wrapper-shape-node.jsx
