+++ canvas_modules/harness/src/client/components/sidepanel/canvas/sidepanel-canvas.jsx
@@ -64,7 +64,7 @@ import {
 	EXAMPLE_APP_LOGIC,
 	EXAMPLE_APP_READ_ONLY,
 	EXAMPLE_APP_PROGRESS,
-	EXAMPLE_APP_REACT_NODES,
+	EXAMPLE_APP_REACT_NODES_CARBON,
 	PALETTE_FLYOUT,
 	PALETTE_MODAL,
 	PALETTE_NONE,
@@ -1132,8 +1132,8 @@ export default class SidePanelForms extends React.Component {
 						labelText={EXAMPLE_APP_TABLES}
 					/>
 					<RadioButton
-						value={EXAMPLE_APP_REACT_NODES}
-						labelText={EXAMPLE_APP_REACT_NODES}
+						value={EXAMPLE_APP_REACT_NODES_CARBON}
+						labelText={EXAMPLE_APP_REACT_NODES_CARBON}
 					/>
 					<RadioButton
 						value={EXAMPLE_APP_NONE}
