+++ canvas_modules/harness/src/client/constants/constants.js
@@ -83,7 +83,7 @@ _defineConstant("EXAMPLE_APP_EXPLAIN2", "Explain2");
 _defineConstant("EXAMPLE_APP_STREAMS", "Streams");
 _defineConstant("EXAMPLE_APP_TABLES", "Tables");
 _defineConstant("EXAMPLE_APP_LOGIC", "Logic");
-_defineConstant("EXAMPLE_APP_REACT_NODES", "React Nodes");
+_defineConstant("EXAMPLE_APP_REACT_NODES_CARBON", "React Nodes - Carbon Charts");
 
 _defineConstant("PALETTE_FLYOUT", "Flyout");
 _defineConstant("PALETTE_MODAL", "Modal");
