+++ canvas_modules/harness/src/styles/custom-canvases.scss
@@ -24,4 +24,4 @@
 @import "../client/components/custom-canvases/explain2/explain2.scss";
 @import "../client/components/custom-canvases/streams/streams.scss";
 @import "../client/components/custom-canvases/tables/tables.scss";
-@import "../client/components/custom-canvases/react-nodes/react-nodes.scss";
+@import "../client/components/custom-canvases/react-nodes-carbon/react-nodes-carbon.scss";
